fix/탑바에서 헤더와 서치바 분리
